payload definition

+++ src/main/java/com/mtoliv/iot/server/message/payLoad/PayloadObjectTypeFlag.java
@@ -2,8 +2,8 @@ package com.mtoliv.iot.server.message.payLoad;
 
 public interface PayloadObjectTypeFlag {
     // 上行
-    int XITONG_ZHUANGTAI = 1 ;
+    byte XITONG_ZHUANGTAI = (byte)1 ;
 
     // 下行
-    int XINXI_CHUANSHU_ZHUANGZHI_SHIJIANG = 90 ;
+    byte XINXI_CHUANSHU_ZHUANGZHI_SHIJIANG = (byte)90 ;
 }
